feat: Add Docker meta information to build workflow
This commit adds a new step to the build workflow that includes Docker meta information. The `docker/metadata-action` is used to set image tags, labels, and maintainers for the `dotwee/latexdevcontainer` image. This change allows for better organization and identification of the image.

The implications of this code change are that it improves the visibility and discoverability of the `dotwee/latexdevcontainer` image by providing relevant metadata. It also ensures that the image is properly labeled with maintainer information and follows Open Containers Image specifications.

These changes can be useful for users who want to easily identify and track versions of the `dotwee/latexdevcontainer` image, as well as for maintaining proper attribution and documentation within the container ecosystem.